+++ src/Release.php
@@ -16,7 +16,7 @@ class Release
     /**
      * Version of the SDK
      */
-    const VERSION = '1.8.1010';
+    const VERSION = '1.8.1011';
 
     /**
      * @param Event $event